Moved query function to action file
